organize React pages and components
Keep route construction separate from page-level data fetching and
reusable presentation components.
