--- src/ui.tsx
@@ -1,53 +1,0 @@
-import { useQuery } from '@tanstack/react-query'
-import { Link, useMatchRoute } from '@tanstack/react-router'
-import type { ReactNode } from 'react'
-import { api } from './api'
-import { DEFAULT_PER_PAGE } from './constants'
-
-export function Layout({ children }: { children: ReactNode }) {
-  const config = useQuery({ queryKey: ['config'], queryFn: api.config })
-  const matchRoute = useMatchRoute()
-  return <>
-    <div className="header">
-      <div className="folders">{config.data?.folders.map((folder) => {
-        const selected = Boolean(matchRoute({ to: '/$folder', params: { folder }, fuzzy: true }))
-        return <span key={folder} className={selected ? 'folder_sel' : 'folder'}>
-          <Link to="/$folder" params={{ folder }} search={{ page: 0, perPage: DEFAULT_PER_PAGE }}>{selected ? `[${folder}]` : folder}</Link>{' '}
-        </span>
-      })}</div>
-      <div className="manual"><a href="https://inside.kmc.gr.jp/rubwiki/%E8%A7%A3%E8%AA%AC%E8%A8%98%E4%BA%8B/%E3%83%8D%E3%83%83%E3%83%88%E3%83%AF%E3%83%BC%E3%82%AF/%E3%83%A1%E3%83%BC%E3%83%AB">解説記事/ネットワーク/メール</a></div>
-    </div>
-    <hr />
-    {children}
-    <hr />
-    <div className="footer">
-      Generated by mail2www<br />
-      Powered by <a href="http://www.ruby-lang.org/">Ruby</a> {config.data?.ruby_version}
-    </div>
-  </>
-}
-
-export function Loading() { return <div className="main">Loading…</div> }
-export function PageError({ error }: { error: Error }) { return <div className="main">{error.message}</div> }
-
-export function Pagination({ folder, page, pages, perPage }: { folder: string, page: number, pages: number, perPage: number }) {
-  return <>{Array.from({ length: pages }, (_, index) => <span key={index} className={index === page ? 'page_sel' : 'page'}>
-    <Link to="/$folder" params={{ folder }} search={{ page: index, perPage }}>{index === page ? `[${index}]` : index}</Link>{' '}
-  </span>)}</>
-}
-
-const urlPattern = /https?:\/\/[^\s<>]+/g
-export function MessageBody({ text, dangerous }: { text: string, dangerous: boolean }) {
-  const chunks: ReactNode[] = []
-  let offset = 0
-  for (const match of text.matchAll(urlPattern)) {
-    const index = match.index
-    chunks.push(text.slice(offset, index))
-    chunks.push(<a key={index} href={match[0]} rel="noreferrer" onClick={(event) => {
-      if (dangerous && !window.confirm('このリンクは危険な可能性があります。本当に開きますか？')) event.preventDefault()
-    }}>{match[0]}</a>)
-    offset = index + match[0].length
-  }
-  chunks.push(text.slice(offset))
-  return <pre className="mail-body">{chunks}</pre>
-}
